mvc refactoring
